moved net.go to proxy/ from proxy/envoy
