+++ proxy/envoy/BUILD
@@ -11,7 +11,6 @@ go_library(
         "header.go",
         "ingress.go",
         "mixer.go",
-        "net.go",
         "policy.go",
         "resolve.go",
         "resources.go",
