+++ proxy/net.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package envoy
+package proxy
 
 import (
 	"net"
